Moves client tests to new package for integration tests
These are part of the CI build
