+++ stravalib/tests/integration/conftest.py
@@ -1,6 +1,7 @@
 import pytest
 
-from stravalib.tests.unit.strava_api_stub import StravaAPIMock
+from stravalib import Client
+from stravalib.tests.integration.strava_api_stub import StravaAPIMock
 
 
 @pytest.fixture
@@ -8,3 +9,8 @@ def mock_strava_api():
     with StravaAPIMock() as api_mock:
         api_mock.add_passthru('https://developers.strava.com/swagger')
         yield api_mock
+
+
+@pytest.fixture
+def client():
+    return Client()
