+++ stravalib/tests/integration/strava_api_stub.py
