+++ stravalib/tests/integration/test_client.py
